refactor: proof logic (#107)
